Arranged classes into packages.
In this interation, the classes that are being used for the whole
project have been moved and organized into seperate packages, and the
default package is only containing the GUI adventure as of right now.
This way, it is easier to navigate the project and find what you're
looking for.

Also renamed Introduction.java to ChapterZero.java